Optionally secure calls with an API key, rename SessionId to Hub
Providing a straightforward mechanism to secure the connections
via an API key should be easy. With this change, users can configure
an API_KEY setting environment variable (or setting in the Azure portal)
and that will automatically require clients to send a matching value.

For additional security, *HTTPS Only* should also be turned on.

The `X-SessionId` was renamed to `X-HUB` since that naming is
already common in SignalR. Unlike SignalR, the hubs in our case
can be any arbitrary string. All clients passing the same value will
share the same "hub" and be able to broadcast messages among
them.