Properly exit the console with error code
This change allows the console to exist cleanly when the
web app failed to start if the ports aren't available.

We don't implement all of the Ctrl+C/Close behavior as
explained in https://github.com/dotnet/coreclr/issues/8565#issuecomment-435698397
because simply canceling the token is enough for our
purposes.